+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/rules/RulesFactory.java
@@ -1,0 +1,6 @@
+package org.openprovenance.prov.rules;
+
+public interface RulesFactory {
+    org.openprovenance.prov.rules.Rules newRules();
+}
+