+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/rules/SimpleRulesFactory.java
@@ -1,0 +1,10 @@
+package org.openprovenance.prov.rules;
+
+public class SimpleRulesFactory implements RulesFactory {
+    public SimpleRulesFactory() {
+    }
+
+    public Rules newRules() {
+        return new Rules();
+    }
+}